Display error if CSV you're deleting is already deleted or it's not your CSV.

+++ app/controllers/api/v0/csv_reports_controller.rb
@@ -119,7 +119,11 @@ class API::V0::CsvReportsController < API::V0::BaseController
   # DELETE /api/v0/csv_reports/:id
 
   def destroy
-    @csv = @current_user.csv_reports.find(params[:id])
+    @csv = @current_user.csv_reports.find_by_id(params[:id])
+    if @csv.blank?
+      raise API::V0::Error.new("CSV ya eliminado o no es tu CSV", 422) and return
+    end
+
     if @csv.destroy
       render :json => {:reload => true}, :status => 200 and return
     else
